Volunteer add pet

+++ backend/src/PetHelpers.Application/Species/Delete/HardDeleteSpeciesHandler.cs
@@ -1,5 +1,7 @@
 ﻿using CSharpFunctionalExtensions;
 using Microsoft.Extensions.Logging;
+using PetHelpers.Application.Database;
+using PetHelpers.Application.Extensions;
 using PetHelpers.Domain.Shared;
 
 namespace PetHelpers.Application.Species.Delete;
@@ -8,27 +10,40 @@ public class HardDeleteSpeciesHandler
 {
     private readonly ISpeciesRepository _repository;
     private readonly ILogger<HardDeleteSpeciesHandler> _logger;
+    private readonly IUnitOfWork _unitOfWork;
+    private readonly DeleteSpeciesCommandValidator _validator;
 
     public HardDeleteSpeciesHandler(
         ISpeciesRepository repository,
-        ILogger<HardDeleteSpeciesHandler> logger)
+        ILogger<HardDeleteSpeciesHandler> logger,
+        IUnitOfWork unitOfWork,
+        DeleteSpeciesCommandValidator validator)
     {
         _repository = repository;
         _logger = logger;
+        _unitOfWork = unitOfWork;
+        _validator = validator;
     }
 
-    public async Task<Result<Guid, Error>> Handle(
-        DeleteSpeciesRequest request,
+    public async Task<Result<Guid, ErrorList>> Handle(
+        DeleteSpeciesCommand command,
         CancellationToken cancellationToken)
     {
-        var speciesResult = await _repository.GetById(request.SpeciesId, cancellationToken);
+        var validationResult = await _validator.ValidateAsync(command, cancellationToken);
+
+        if (validationResult.IsValid == false)
+            return validationResult.ToList();
+
+        var speciesResult = await _repository.GetById(command.SpeciesId, cancellationToken);
 
         if (speciesResult.IsFailure)
-            return speciesResult.Error;
+            return speciesResult.Error.ToErrorList();
 
         var species = speciesResult.Value;
 
-        var result = await _repository.Delete(species, cancellationToken);
+        var result = _repository.Delete(species, cancellationToken);
+
+        await _unitOfWork.SaveChanges(cancellationToken);
 
         _logger.LogInformation("Deleted species with id: {speciesId}", result);
